Added support for populating dirs owned by named users/groups
This is now possible with:

```
populate:
  method: config
  entries:
    - path: /var/lib/database
      type: directory
      mode: "755"
      user: dbuser
      group: dbgroup
```

While it was possible already to chown directories with `uid` and `gid`,
those only support numeric IDs, and those might not always be known.

`uid` and `gid` are still supported, while `user` and `group` are introduced,
and are consistent with the mount_point `chown` attribute, and accept both
numeric and named users and groups.